Refactored: to add addtional validations and used implementaion of BeanUtils

+++ src/main/java/com/bridgelabz/addressbookapp/repository/AddressBookRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.bridgelabz.addressbookapp.model.AddressBookData;
 
 @Repository
-public interface AddressBookRepository extends CrudRepository<AddressBookData, Integer> {
+public interface AddressBookRepository extends CrudRepository<AddressBookData, Long> {
 
 }
